api-server/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-filter==25.1
djangorestframework==3.16.1
sqlparse==0.5.3
json-log-formatter==1.1.1
gunicorn==21.2.0
psycopg[binary]==3.2.13
coverage==7.10.6
